# CUDA: use mma FA kernel for gqa > 4 on RTX 4000